return np.array(zip_res, dtype=np.int32)

+++ zip_array.py
@@ -55,14 +55,14 @@ def zip_array(array, data=0.):  # 压缩成稀疏数组
         for j in range(len(array[0])):
             if array[i][j] != data:
                 zip_res.append([i, j, array[i][j]])
-    return np.array(zip_res)
+    return np.array(zip_res, dtype=np.int32)
 
 
 def recovery_array(array, data=0.):  # 恢复数组
     recovery_res = []
-    for i in range(int(array[0][0])):
-        recovery_res.append([data for i in range(int(array[0][1]))])
+    for i in range(array[0][0]):
+        recovery_res.append([data for i in range(array[0][1])])
     for i in range(1, len(array)):
         # print(len(recovery_res[0]))
-        recovery_res[int(array[i][0])][int(array[i][1])] = array[i][2]
+        recovery_res[array[i][0]][array[i][1]] = array[i][2]
     return np.array(recovery_res)
